Added null check for ID in onDelete

+++ app/lib/controllers/log_ui_controller.dart
@@ -93,6 +93,11 @@ class LogUIController extends ChangeNotifier {
 
             Navigator.pop(sheetContext);
 
+            if (log.id == null) {
+              _showError(context, "Cannot delete unsaved log"); //TODO: localisation 'cannotDeleteUnsavedLogMessage'
+              return;
+            }
+
             await logService.deleteLog(log.id!);
 
             await periodService.refreshData(
